Language.EO.Phi

Documentation

normalize :: Program -> Program

Normalize an input 𝜑-program.

parseProgram :: String -> Either String Program

Parse a Program or return a parsing error.

unsafeParseObject :: String -> Object

Parse an Object from a String. May throw an error if input has a syntactical or lexical errors.

unsafeParseProgram :: String -> Program

Parse a Program from a String. May throw an error if input has a syntactical or lexical errors.

unsafeParseProgramFromFile :: FilePath -> IO Program

module Language.EO.Phi.Syntax