Language.EO.Phi.Rules.Common

Documentation

>>> :set -XOverloadedStrings
>>> :set -XOverloadedLists
>>> import Language.EO.Phi.Syntax

type EvaluationState = ()

State of evaluation is not needed yet, but it might be in the future

type NamedRule = (String, Rule)

type Atoms = HashMap String (String -> Object -> EvaluationState -> DataizeChain (Object, EvaluationState))

data Context

Constructors

Context

Fields

builtinRules :: Bool

allRules :: [NamedRule]

enabledAtoms :: Atoms

knownAtoms :: Atoms

outerFormations :: NonEmpty Object

currentAttr :: Attribute

insideFormation :: Bool

Temporary hack for applying Ksi and Phi rules when dataizing

insideAbstractFormation :: Bool

dataizePackage :: Bool

Temporary flag to only dataize Package attributes for the top-level formation.

minimizeTerms :: Bool

insideSubObject :: Bool

sameContext :: Context -> Context -> Bool

type Rule = Context -> Object -> [Object]

A rule tries to apply a transformation to the root object, if possible.

applyOneRuleAtRoot :: Context -> Object -> [(String, Object)]

extendContextWith :: Object -> Context -> Context

isEmptyBinding :: Binding -> Bool

withSubObject :: (Context -> Object -> [(String, Object)]) -> Context -> Object -> [(String, Object)]

propagateName1 :: (a -> b) -> (name, a) -> (name, b)

Given a unary function that operates only on plain objects, converts it to a function that operates on named objects

propagateName2 :: (a -> b -> c) -> (name, a) -> b -> (name, c)

Given a binary function that operates only on plain objects, converts it to a function that operates on named objects

withSubObjectBindings :: (Context -> Object -> [(String, Object)]) -> Context -> [Binding] -> [(String, [Binding])]

withSubObjectBinding :: (Context -> Object -> [(String, Object)]) -> Context -> Binding -> [(String, Binding)]

applyOneRule :: Context -> Object -> [(String, Object)]

isNF :: Context -> Object -> Bool

applyRules :: Context -> Object -> [Object]

Apply rules until we get a normal form.

data ApplicationLimits

Constructors

ApplicationLimits

Fields

maxDepth :: Int

maxTermSize :: Int

defaultApplicationLimits :: Int -> ApplicationLimits

objectSize :: Object -> Int

bindingSize :: Binding -> Int

applyRulesWith :: ApplicationLimits -> Context -> Object -> [Object]

A variant of applyRules with a maximum application depth.

equalProgram :: Program -> Program -> Bool

equalObject :: Object -> Object -> Bool

equalObjectNamed :: (String, Object) -> (String, Object) -> Bool

equalBindings :: [Binding] -> [Binding] -> Bool

equalBinding :: Binding -> Binding -> Bool

Chain variants

data LogEntry log

Constructors

LogEntry

Fields

logEntryMessage :: String

logEntryLog :: log

logEntryLevel :: Int

newtype Chain log result

Constructors

Chain

Fields

runChain :: Context -> [([LogEntry log], result)]

type NormalizeChain = Chain Object

type DataizeChain = Chain (Either Object Bytes)

logStep :: String -> info -> Chain info ()

incLogLevel :: Chain info a -> Chain info a

choose :: [a] -> Chain log a

msplit :: Chain log a -> Chain log (Maybe (a, Chain log a))

transformLogs :: (log1 -> log2) -> Chain log1 a -> Chain log2 a

transformNormLogs :: NormalizeChain a -> DataizeChain a

listen :: Chain log a -> Chain log (a, [LogEntry log])

minimizeObject' :: DataizeChain (Either Object Bytes) -> DataizeChain (Either Object Bytes)

minimizeObject :: DataizeChain (Either Object Bytes) -> DataizeChain (Either Object Bytes)

smallestObject :: [Either Object bytes] -> Object -> Object

getContext :: Chain a Context

withContext :: Context -> Chain log a -> Chain log a

modifyContext :: (Context -> Context) -> Chain log a -> Chain log a

applyRulesChain' :: Context -> Object -> [([LogEntry Object], Object)]

applyRulesChain :: Object -> NormalizeChain Object

Apply the rules until the object is normalized, preserving the history (chain) of applications.

applyRulesChainWith' :: ApplicationLimits -> Context -> Object -> [([LogEntry Object], Object)]

applyRulesChainWith :: ApplicationLimits -> Object -> NormalizeChain Object

A variant of applyRulesChain with a maximum application depth.

Helpers

lookupBinding :: Attribute -> [Binding] -> Maybe Object

Lookup a binding by the attribute name.

objectBindings :: Object -> [Binding]

isRhoBinding :: Binding -> Bool

hideRhoInBinding :: Binding -> Binding

hideRho :: Object -> Object

hideRhoInBinding1 :: Binding -> Binding

hideRho1 :: Object -> Object