Language.EO.Phi.Syntax

Documentation

module Language.EO.Phi.Syntax.Abs

desugar :: Object -> Object

printTree :: (Pretty a, SugarableFinally a) => a -> String

The top-level printing method.

printTreeDontSugar :: Pretty a => a -> String

Conversion to Bytes

intToBytes :: Int -> Bytes

Convert an Int into Bytes representation.

>>> intToBytes 7
Bytes "00-00-00-00-00-00-00-07"
>>> intToBytes (3^33)
Bytes "00-13-BF-EF-A6-5A-BB-83"
>>> intToBytes (-1)
Bytes "FF-FF-FF-FF-FF-FF-FF-FF"

int64ToBytes :: Int64 -> Bytes

Convert an Int64 into Bytes representation.

>>> int64ToBytes 7
Bytes "00-00-00-00-00-00-00-07"
>>> int64ToBytes (3^33)
Bytes "00-13-BF-EF-A6-5A-BB-83"
>>> int64ToBytes (-1)
Bytes "FF-FF-FF-FF-FF-FF-FF-FF"

int32ToBytes :: Int32 -> Bytes

Convert an Int32 into Bytes representation.

>>> int32ToBytes 7
Bytes "00-00-00-07"
>>> int32ToBytes (3^33)
Bytes "A6-5A-BB-83"
>>> int32ToBytes (-1)
Bytes "FF-FF-FF-FF"

int16ToBytes :: Int16 -> Bytes

Convert an Int16 into Bytes representation.

>>> int16ToBytes 7
Bytes "00-07"
>>> int16ToBytes (3^33)
Bytes "BB-83"
>>> int16ToBytes (-1)
Bytes "FF-FF"

floatToBytes :: Double -> Bytes

Encode Double as Bytes following IEEE754.

Note: it is called "float" in EO, but it actually occupies 8 bytes so it corresponds to Double.

>>> floatToBytes 0
Bytes "00-00-00-00-00-00-00-00"

>>> floatToBytes (-0.1)
Bytes "BF-B9-99-99-99-99-99-9A"

>>> floatToBytes (1/0)       -- Infinity
Bytes "7F-F0-00-00-00-00-00-00"

>>> floatToBytes (asin 2) `elem` ["FF-F8-00-00-00-00-00-00", "7F-F8-00-00-00-00-00-00"]  -- sNaN or qNaN
True

boolToBytes :: Bool -> Bytes

Convert Bool to Bytes.

>>> boolToBytes False
Bytes "00-"
>>> boolToBytes True
Bytes "01-"

stringToBytes :: String -> Bytes

Encode String as Bytes.

>>> stringToBytes "Hello, world!"
Bytes "48-65-6C-6C-6F-2C-20-77-6F-72-6C-64-21"

>>> stringToBytes "Привет, мир!"
Bytes "D0-9F-D1-80-D0-B8-D0-B2-D0-B5-D1-82-2C-20-D0-BC-D0-B8-D1-80-21"

>>> stringToBytes  "hello, 大家!"
Bytes "68-65-6C-6C-6F-2C-20-E5-A4-A7-E5-AE-B6-21"

Conversion from Bytes

bytesToInt :: Bytes -> Int

Parse Bytes as Int.

>>> bytesToInt "00-13-BF-EF-A6-5A-BB-83"
5559060566555523
>>> bytesToInt "AB-"
171

May error on invalid Bytes:

>>> bytesToInt "s"
*** Exception: Prelude.head: empty list
...
...
...
...
...
...

bytesToInt64 :: Bytes -> Int64

Parse Bytes as Int64.

>>> bytesToInt64 "00-13-BF-EF-A6-5A-BB-83"
5559060566555523
>>> bytesToInt64 "AB-"
171

May error on invalid Bytes:

>>> bytesToInt64 "s"
*** Exception: Prelude.head: empty list
...
...
...
...
...
...

bytesToInt32 :: Bytes -> Int32

Parse Bytes as Int32.

>>> bytesToInt32 "A6-5A-BB-83"
-1504003197
>>> bytesToInt32 "AB-"
171

May error on invalid Bytes:

>>> bytesToInt32 "s"
*** Exception: Prelude.head: empty list
...
...
...
...
...
...

bytesToInt16 :: Bytes -> Int16

Parse Bytes as Int16.

>>> bytesToInt16 "BB-83"
-17533
>>> bytesToInt16 "AB-"
171

May error on invalid Bytes:

>>> bytesToInt16 "s"
*** Exception: Prelude.head: empty list
...
...
...
...
...
...

bytesToFloat :: Bytes -> Double

Decode Double from Bytes following IEEE754.

>>> bytesToFloat "00-00-00-00-00-00-00-00"
0.0

>>> bytesToFloat "BF-B9-99-99-99-99-99-9A"
-0.1

>>> bytesToFloat "7F-F0-00-00-00-00-00-00"
Infinity

>>> bytesToFloat "FF-F8-00-00-00-00-00-00"
NaN

bytesToString :: Bytes -> String

Decode String from Bytes.

>>> bytesToString "48-65-6C-6C-6F-2C-20-77-6F-72-6C-64-21"
"Hello, world!"

bytesToBool :: Bytes -> Bool

Interpret Bytes as Bool.

Zero is interpreted as False.

>>> bytesToBool "00-"
False

>>> bytesToBool "00-00"
False

Everything else is interpreted as True.

>>> bytesToBool "01-"
True

>>> bytesToBool "00-01"
True

>>> bytesToBool "AB-CD"
True

Wrapping Bytes into Object

wrapBytesInConstInt :: Bytes -> Object

wrapBytesInConstInt64 :: Bytes -> Object

wrapBytesInConstInt32 :: Bytes -> Object

wrapBytesInConstInt16 :: Bytes -> Object

wrapBytesInConstFloat :: Bytes -> Object

wrapBytesInConstString :: Bytes -> Object

wrapBytesInBytes :: Bytes -> Object

wrapBytesInInt :: Bytes -> Object

wrapBytesInFloat :: Bytes -> Object

wrapBytesAsBool :: Bytes -> Object

wrapBytesInString :: Bytes -> Object

wrapTermination :: Object

Functions over Bytes

sliceBytes :: Bytes -> Int -> Int -> Bytes

Select a slice (section) of Bytes.

>>> sliceBytes "12-34-56" 1 1
Bytes "34-"

>>> sliceBytes "12-34-56" 1 0
Bytes "00-"

>>> sliceBytes "12-34-56" 0 2
Bytes "12-34"

concatBytes :: Bytes -> Bytes -> Bytes

Concatenate Bytes. FIXME: we should really use ByteString instead of the underlying String representation.

>>> concatBytes "00-" "01-02"
Bytes "00-01-02"

>>> concatBytes "03-04" "01-02"
Bytes "03-04-01-02"

>>> concatBytes "03-04" "01-"
Bytes "03-04-01"

Helpers

chunksOf :: Int -> [a] -> [[a]]

Split a list into chunks of given size. All lists in the result are guaranteed to have length less than or equal to the given size.

>>> chunksOf 2 "012345678"
["01","23","45","67","8"]

See paddedLeftChunksOf for a version with padding to guarantee exact chunk size.

paddedLeftChunksOf :: a -> Int -> [a] -> [[a]]

Split a list into chunks of given size, padding on the left if necessary. All lists in the result are guaranteed to have given size.

>>> paddedLeftChunksOf '0' 2 "1234567"
["01","23","45","67"]
>>> paddedLeftChunksOf '0' 2 "123456"
["12","34","56"]

n > 0  ==>  all (\chunk -> length chunk == n) (paddedLeftChunksOf c n s)

normalizeBytes :: String -> String

Normalize the bytestring representation to fit valid Bytes token.

>>> normalizeBytes "238714ABCDEF"
"23-87-14-AB-CD-EF"

>>> normalizeBytes "0238714ABCDEF"
"00-23-87-14-AB-CD-EF"

>>> normalizeBytes "4"
"04-"

parseWith :: (DesugarableInitially a, CheckableSyntaxInitially a) => ([Token] -> Either String a) -> String -> Either String a

errorExpectedDesugaredObject :: Object -> a

errorExpectedDesugaredBinding :: Binding -> a

errorExpectedDesugaredAttribute :: Attribute -> a

Classes

class SugarableFinally a where

Minimal complete definition

Nothing

Methods

sugarFinally :: a -> a

Pattern synonyms

pattern AlphaBinding' :: Attribute -> Object -> Binding

pattern AlphaBinding'' :: LabelId -> [Attribute] -> Object -> Binding

Orphan instances

IsString Attribute

Instance details

Methods

fromString :: String -> Attribute

IsString AttributeSugar

Instance details

Methods

fromString :: String -> AttributeSugar

IsString Binding

Instance details

Methods

fromString :: String -> Binding

IsString MetaId

Instance details

Methods

fromString :: String -> MetaId

IsString Object

Instance details

Methods

fromString :: String -> Object

IsString ObjectHead

Instance details

Methods

fromString :: String -> ObjectHead

IsString PeeledObject

Instance details

Methods

fromString :: String -> PeeledObject

IsString Program

Instance details

Methods

fromString :: String -> Program

IsString RuleAttribute

Instance details

Methods

fromString :: String -> RuleAttribute