Language.EO.Phi.Dataize

Documentation

dataizeStep :: Context -> Object -> (Context, Either Object Bytes)

Perform one step of dataization to the object (if possible).

dataizeStep' :: Context -> Object -> Either Object Bytes

dataizeRecursively :: Context -> Object -> Either Object Bytes

Recursively perform normalization and dataization until we get bytes in the end.

dataizeStepChain' :: Context -> Object -> ([LogEntry (Either Object Bytes)], Either Object Bytes)

dataizeStepChain :: Object -> DataizeChain (Context, Either Object Bytes)

Perform one step of dataization to the object (if possible), reporting back individiual steps.

dataizeRecursivelyChain' :: Context -> Object -> ([LogEntry (Either Object Bytes)], Either Object Bytes)

dataizeRecursivelyChain :: Bool -> Object -> DataizeChain (Either Object Bytes)

Recursively perform normalization and dataization until we get bytes in the end, reporting intermediate steps

evaluateDataizationFunChain

Arguments

:: (res -> Bytes)
How to convert the result back to bytes

-> (Bytes -> a)
How to interpret the bytes in terms of the given data type

-> (Bytes -> Object)
How to wrap the bytes in an object

-> (a -> a -> res)
A binary function on the data

-> Object

-> EvaluationState

-> DataizeChain (Object, EvaluationState)

Given converters between Bytes and some data type, a binary function on this data type, an object, and the current state of evaluation, returns the new object and a possibly modified state along with intermediate steps.

evaluateBinaryDataizationFunChain

Arguments

:: (res -> Bytes)
How to convert the result back to bytes

-> (Bytes -> a)
How to interpret the bytes in terms of the given data type

-> (Bytes -> Object)
How to wrap the bytes in an object

-> (Object -> Object)
Extract the 1st argument to be dataized

-> (Object -> Object)
Extract the 2nd argument to be dataized

-> (a -> a -> res)
A binary function on the argument

-> String
Name of the atom.

-> Object

-> EvaluationState

-> DataizeChain (Object, EvaluationState)

evaluateUnaryDataizationFunChain

Arguments

:: (res -> Bytes)
How to convert the result back to bytes

-> (Bytes -> a)
How to interpret the bytes in terms of the given data type

-> (Bytes -> Object)
How to wrap the bytes in an object

-> (Object -> Object)
Extract the argument to be dataized

-> (a -> res)
A unary function on the argument

-> String

-> Object

-> EvaluationState

-> DataizeChain (Object, EvaluationState)

Unary functions operate on the given object without any additional parameters

evaluateIntIntIntFunChain :: (Int -> Int -> Int) -> String -> Object -> EvaluationState -> DataizeChain (Object, EvaluationState)

evaluateIntIntBoolFunChain :: (Int -> Int -> Bool) -> String -> Object -> EvaluationState -> DataizeChain (Object, EvaluationState)

evaluateBytesBytesBytesFunChain :: (Int -> Int -> Int) -> String -> Object -> EvaluationState -> DataizeChain (Object, EvaluationState)

evaluateBytesBytesFunChain :: (Int -> Int) -> String -> Object -> EvaluationState -> DataizeChain (Object, EvaluationState)

evaluateFloatFloatFloatFunChain :: (Double -> Double -> Double) -> String -> Object -> EvaluationState -> DataizeChain (Object, EvaluationState)

evaluateBuiltinFunChain :: String -> Object -> EvaluationState -> DataizeChain (Object, EvaluationState)

Like evaluateDataizationFunChain but specifically for the built-in functions. This function is not safe. It returns undefined for unknown functions

evaluateBuiltinFunChainUnknown :: String -> Object -> EvaluationState -> DataizeChain (Object, EvaluationState)

evaluateBuiltinFun :: Context -> String -> Object -> EvaluationState -> (Object, EvaluationState)

Like evaluateDataizationFun but specifically for the built-in functions. This function is not safe. It returns undefined for unknown functions

evaluateIODataizationFunChain :: IO String -> Object -> EvaluationState -> DataizeChain (Object, EvaluationState)

extractRho :: Object -> Object

extractAlpha0 :: Object -> Object

extractLabel :: String -> Object -> Object

wrapBytesInInt :: Bytes -> Object

wrapBytesInFloat :: Bytes -> Object

wrapBytesInString :: Bytes -> Object

wrapBytesInBytes :: Bytes -> Object

wrapTermination :: Object

wrapBytesAsBool :: Bytes -> Object