Language.EO.Phi.Dataize

Documentation

dataizeStep :: Context -> Object -> Either Object Bytes

Perform one step of dataization to the object (if possible).

dataizeRecursively :: Context -> Object -> Either Object Bytes

Recursively perform normalization and dataization until we get bytes in the end.

dataizeStepChain :: Context -> Object -> [(String, Either Object Bytes)]

Perform one step of dataization to the object (if possible), reporting back individiual steps.

dataizeRecursivelyChain :: Context -> Object -> [(String, Either Object Bytes)]

Recursively perform normalization and dataization until we get bytes in the end, reporting intermediate steps