Language.EO.Phi.Rules.Common

Documentation

>>> :set -XOverloadedStrings
>>> :set -XOverloadedLists
>>> import Language.EO.Phi.Syntax

parseWith :: ([Token] -> Either String a) -> String -> Either String a

unsafeParseWith :: ([Token] -> Either String a) -> String -> a

Parse a Object from a String. May throw an error if input has a syntactical or lexical errors.

type EvaluationState = ()

State of evaluation is not needed yet, but it might be in the future

type NamedRule = (String, Rule)

type Atoms = HashMap String (String -> Object -> EvaluationState -> DataizeChain (Object, EvaluationState))

data Context

Constructors

Context

Fields

builtinRules :: Bool

allRules :: [NamedRule]

enabledAtoms :: Atoms

knownAtoms :: Atoms

outerFormations :: NonEmpty Object

currentAttr :: Attribute

insideFormation :: Bool

Temporary hack for applying Ksi and Phi rules when dataizing

insideAbstractFormation :: Bool

dataizePackage :: Bool

Temporary flag to only dataize Package attributes for the top-level formation.

minimizeTerms :: Bool

insideSubObject :: Bool

sameContext :: Context -> Context -> Bool

type Rule = Context -> Object -> [Object]

A rule tries to apply a transformation to the root object, if possible.

applyOneRuleAtRoot :: Context -> Object -> [(String, Object)]

extendContextWith :: Object -> Context -> Context

isEmptyBinding :: Binding -> Bool

withSubObject :: (Context -> Object -> [(String, Object)]) -> Context -> Object -> [(String, Object)]

propagateName1 :: (a -> b) -> (name, a) -> (name, b)

Given a unary function that operates only on plain objects, converts it to a function that operates on named objects

propagateName2 :: (a -> b -> c) -> (name, a) -> b -> (name, c)

Given a binary function that operates only on plain objects, converts it to a function that operates on named objects

withSubObjectBindings :: (Context -> Object -> [(String, Object)]) -> Context -> [Binding] -> [(String, [Binding])]

withSubObjectBinding :: (Context -> Object -> [(String, Object)]) -> Context -> Binding -> [(String, Binding)]

applyOneRule :: Context -> Object -> [(String, Object)]

isNF :: Context -> Object -> Bool

applyRules :: Context -> Object -> [Object]

Apply rules until we get a normal form.

data ApplicationLimits

Constructors

ApplicationLimits

Fields

maxDepth :: Int

maxTermSize :: Int

defaultApplicationLimits :: Int -> ApplicationLimits

objectSize :: Object -> Int

bindingSize :: Binding -> Int

applyRulesWith :: ApplicationLimits -> Context -> Object -> [Object]

A variant of applyRules with a maximum application depth.

equalProgram :: Program -> Program -> Bool

equalObject :: Object -> Object -> Bool

equalObjectNamed :: (String, Object) -> (String, Object) -> Bool

equalBindings :: [Binding] -> [Binding] -> Bool

equalBinding :: Binding -> Binding -> Bool

Chain variants

data LogEntry log

Constructors

LogEntry

Fields

logEntryMessage :: String

logEntryLog :: log

logEntryLevel :: Int

newtype Chain log result

Constructors

Chain

Fields

runChain :: Context -> [([LogEntry log], result)]

type NormalizeChain = Chain Object

type DataizeChain = Chain (Either Object Bytes)

logStep :: String -> info -> Chain info ()

incLogLevel :: Chain info a -> Chain info a

choose :: [a] -> Chain log a

msplit :: Chain log a -> Chain log (Maybe (a, Chain log a))

transformLogs :: (log1 -> log2) -> Chain log1 a -> Chain log2 a

transformNormLogs :: NormalizeChain a -> DataizeChain a

listen :: Chain log a -> Chain log (a, [LogEntry log])

minimizeObject' :: DataizeChain (Either Object Bytes) -> DataizeChain (Either Object Bytes)

minimizeObject :: DataizeChain (Either Object Bytes) -> DataizeChain (Either Object Bytes)

smallestObject :: [Either Object bytes] -> Object -> Object

getContext :: Chain a Context

withContext :: Context -> Chain log a -> Chain log a

modifyContext :: (Context -> Context) -> Chain log a -> Chain log a

applyRulesChain' :: Context -> Object -> [([LogEntry Object], Object)]

applyRulesChain :: Object -> NormalizeChain Object

Apply the rules until the object is normalized, preserving the history (chain) of applications.

applyRulesChainWith' :: ApplicationLimits -> Context -> Object -> [([LogEntry Object], Object)]

applyRulesChainWith :: ApplicationLimits -> Object -> NormalizeChain Object

A variant of applyRulesChain with a maximum application depth.

Helpers

lookupBinding :: Attribute -> [Binding] -> Maybe Object

Lookup a binding by the attribute name.

objectBindings :: Object -> [Binding]

padLeft :: Int -> [Char] -> [Char]

chunksOf :: Int -> [a] -> [[a]]

Split a list into chunks of given size. All lists in the result are guaranteed to have length less than or equal to the given size.

>>> chunksOf 2 "012345678"
["01","23","45","67","8"]

See paddedLeftChunksOf for a version with padding to guarantee exact chunk size.

paddedLeftChunksOf :: a -> Int -> [a] -> [[a]]

Split a list into chunks of given size, padding on the left if necessary. All lists in the result are guaranteed to have given size.

>>> paddedLeftChunksOf '0' 2 "1234567"
["01","23","45","67"]
>>> paddedLeftChunksOf '0' 2 "123456"
["12","34","56"]

n > 0  ==>  all (\chunk -> length chunk == n) (paddedLeftChunksOf c n s)

normalizeBytes :: String -> String

Normalize the bytestring representation to fit valid Bytes token.

>>> normalizeBytes "238714ABCDEF"
"23-87-14-AB-CD-EF"

>>> normalizeBytes "0238714ABCDEF"
"00-23-87-14-AB-CD-EF"

>>> normalizeBytes "4"
"04-"

concatBytes :: Bytes -> Bytes -> Bytes

Concatenate Bytes. FIXME: we should really use ByteString instead of the underlying String representation.

>>> concatBytes "00-" "01-02"
Bytes "00-01-02"

>>> concatBytes "03-04" "01-02"
Bytes "03-04-01-02"

>>> concatBytes "03-04" "01-"
Bytes "03-04-01"

sliceBytes :: Bytes -> Int -> Int -> Bytes

Select a slice (section) of Bytes.

>>> sliceBytes "12-34-56" 1 1
Bytes "34-"

>>> sliceBytes "12-34-56" 1 0
Bytes "00-"

>>> sliceBytes "12-34-56" 0 2
Bytes "12-34"

intToBytes :: Int -> Bytes

Convert an Int into Bytes representation.

>>> intToBytes 7
Bytes "00-00-00-00-00-00-00-07"
>>> intToBytes (3^33)
Bytes "00-13-BF-EF-A6-5A-BB-83"
>>> intToBytes (-1)
Bytes "FF-FF-FF-FF-FF-FF-FF-FF"

bytesToInt :: Bytes -> Int

Parse Bytes as Int.

>>> bytesToInt "00-13-BF-EF-A6-5A-BB-83"
5559060566555523
>>> bytesToInt "AB-"
171

May error on invalid Bytes:

>>> bytesToInt "s"
*** Exception: Prelude.head: empty list
...
...
...
...
...
...

boolToBytes :: Bool -> Bytes

Convert Bool to Bytes.

>>> boolToBytes False
Bytes "00-"
>>> boolToBytes True
Bytes "01-"

bytesToBool :: Bytes -> Bool

Interpret Bytes as Bool.

Zero is interpreted as False.

>>> bytesToBool "00-"
False

>>> bytesToBool "00-00"
False

Everything else is interpreted as True.

>>> bytesToBool "01-"
True

>>> bytesToBool "00-01"
True

>>> bytesToBool "AB-CD"
True

stringToBytes :: String -> Bytes

Encode String as Bytes.

>>> stringToBytes "Hello, world!"
Bytes "48-65-6C-6C-6F-2C-20-77-6F-72-6C-64-21"

>>> stringToBytes "Привет, мир!"
Bytes "D0-9F-D1-80-D0-B8-D0-B2-D0-B5-D1-82-2C-20-D0-BC-D0-B8-D1-80-21"

>>> stringToBytes  "hello, 大家!"
Bytes "68-65-6C-6C-6F-2C-20-E5-A4-A7-E5-AE-B6-21"

bytestringToBytes :: ByteString -> Bytes

bytesToByteString :: Bytes -> ByteString

bytesToString :: Bytes -> String

Decode String from Bytes.

>>> bytesToString "48-65-6C-6C-6F-2C-20-77-6F-72-6C-64-21"
"Hello, world!"

floatToBytes :: Double -> Bytes

Encode Double as Bytes following IEEE754.

Note: it is called "float" in EO, but it actually occupies 8 bytes so it corresponds to Double.

>>> floatToBytes 0
Bytes "00-00-00-00-00-00-00-00"

>>> floatToBytes (-0.1)
Bytes "BF-B9-99-99-99-99-99-9A"

>>> floatToBytes (1/0)       -- Infinity
Bytes "7F-F0-00-00-00-00-00-00"

>>> floatToBytes (asin 2) `elem` ["FF-F8-00-00-00-00-00-00", "7F-F8-00-00-00-00-00-00"]  -- sNaN or qNaN
True

bytesToFloat :: Bytes -> Double

Decode Double from Bytes following IEEE754.

>>> bytesToFloat "00-00-00-00-00-00-00-00"
0.0

>>> bytesToFloat "BF-B9-99-99-99-99-99-9A"
-0.1

>>> bytesToFloat "7F-F0-00-00-00-00-00-00"
Infinity

>>> bytesToFloat "FF-F8-00-00-00-00-00-00"
NaN

isRhoBinding :: Binding -> Bool

hideRhoInBinding :: Binding -> Binding

hideRho :: Object -> Object

hideRhoInBinding1 :: Binding -> Binding

hideRho1 :: Object -> Object

Orphan instances

IsString Attribute

Instance details

Methods

fromString :: String -> Attribute

IsString Binding

Instance details

Methods

fromString :: String -> Binding

IsString MetaId

Instance details

Methods

fromString :: String -> MetaId

IsString Object

Instance details

Methods

fromString :: String -> Object

IsString ObjectHead

Instance details

Methods

fromString :: String -> ObjectHead

IsString PeeledObject

Instance details

Methods

fromString :: String -> PeeledObject

IsString Program

Instance details

Methods

fromString :: String -> Program

IsString RuleAttribute

Instance details

Methods

fromString :: String -> RuleAttribute