Language.EO.Phi.Rules.Common

Documentation

>>> :set -XOverloadedStrings
>>> :set -XOverloadedLists
>>> import Language.EO.Phi.Syntax

parseWith :: ([Token] -> Either String a) -> String -> Either String a

unsafeParseWith :: ([Token] -> Either String a) -> String -> a

Parse a Object from a String. May throw an error if input has a syntactical or lexical errors.

type NamedRule = (String, Rule)

data Context

Constructors

Context

Fields

allRules :: [NamedRule]

outerFormations :: NonEmpty Object

currentAttr :: Attribute

insideFormation :: Bool

Temporary hack for applying Ksi and Phi rules when dataizing

defaultContext :: [NamedRule] -> Object -> Context

type Rule = Context -> Object -> [Object]

A rule tries to apply a transformation to the root object, if possible.

applyOneRuleAtRoot :: Context -> Object -> [(String, Object)]

extendContextWith :: Object -> Context -> Context

withSubObject :: (Context -> Object -> [(String, Object)]) -> Context -> Object -> [(String, Object)]

propagateName1 :: (a -> b) -> (name, a) -> (name, b)

Given a unary function that operates only on plain objects, converts it to a function that operates on name objects

propagateName2 :: (a -> b -> c) -> (name, a) -> b -> (name, c)

Given a binary function that operates only on plain objects, converts it to a function that operates on name objects

withSubObjectBindings :: (Context -> Object -> [(String, Object)]) -> Context -> [Binding] -> [(String, [Binding])]

withSubObjectBinding :: (Context -> Object -> [(String, Object)]) -> Context -> Binding -> [(String, Binding)]

applyOneRule :: Context -> Object -> [(String, Object)]

isNF :: Context -> Object -> Bool

applyRules :: Context -> Object -> [Object]

Apply rules until we get a normal form.

>>> mapM_ (putStrLn . Language.EO.Phi.printTree) (applyRules (Context [rule6] ["⟦ a ↦ ⟦ b ↦ ⟦ ⟧ ⟧.b ⟧"]) "⟦ a ↦ ⟦ b ↦ ⟦ ⟧ ⟧.b ⟧.a")

data ApplicationLimits

Constructors

ApplicationLimits

Fields

maxDepth :: Int

maxTermSize :: Int

defaultApplicationLimits :: Int -> ApplicationLimits

objectSize :: Object -> Int

bindingSize :: Binding -> Int

applyRulesWith :: ApplicationLimits -> Context -> Object -> [Object]

A variant of applyRules with a maximum application depth.

equalProgram :: Program -> Program -> Bool

equalObject :: Object -> Object -> Bool

equalObjectNamed :: (String, Object) -> (String, Object) -> Bool

equalBindings :: [Binding] -> [Binding] -> Bool

equalBinding :: Binding -> Binding -> Bool

applyRulesChain :: Context -> Object -> [[(String, Object)]]

Apply the rules until the object is normalized, preserving the history (chain) of applications.

applyRulesChainWith :: ApplicationLimits -> Context -> Object -> [[(String, Object)]]

A variant of applyRulesChain with a maximum application depth.

Helpers

lookupBinding :: Attribute -> [Binding] -> Maybe Object

Lookup a binding by the attribute name.

objectBindings :: Object -> [Binding]

intToBytes :: Int -> Bytes

bytesToInt :: Bytes -> Int

Assuming the bytes are well-formed (otherwise crashes)

minNu :: Int

class HasMaxNu a where

Methods

getMaxNu :: a -> Int

get maximum vertex index

>>> getMaxNu @Object "⟦ a ↦ ⟦ ν ↦ ⟦ Δ ⤍ 03- ⟧ ⟧, b ↦ ⟦ ⟧ ⟧"
3

intToBytesObject :: Int -> Object

nuCountAsDataObj :: Object -> Object

Orphan instances

IsString Attribute

Instance details

Methods

fromString :: String -> Attribute

IsString Binding

Instance details

Methods

fromString :: String -> Binding

IsString Object

Instance details

Methods

fromString :: String -> Object

IsString ObjectHead

Instance details

Methods

fromString :: String -> ObjectHead

IsString PeeledObject

Instance details

Methods

fromString :: String -> PeeledObject

IsString Program

Instance details

Methods

fromString :: String -> Program

IsString RuleAttribute

Instance details

Methods

fromString :: String -> RuleAttribute